Disposing stores and cleaning code
